# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "AutoTuneVramRequest": {
        "description": "Request body for ``POST /api/training/auto-tune-vram``.\n\nAll fields are optional. When the hardware fields are omitted the\nbackend queries ``GpuInfo`` for VRAM and ``psutil.virtual_memory()``\nfor host RAM. Supplying them explicitly is the entry point the\nStage F smoke script uses to simulate a smaller card on the 5090.",
        "properties": {
          "profile": {
            "default": "video",
            "description": "Training profile to tune for. The image profile (frames=1) and video profile (121 frames) have different VRAM curves and therefore different tier tables. Defaults to 'video' to match TrainingConfig.profile.",
            "enum": [
              "image",
              "video"
            ],
            "title": "Profile",
            "type": "string"
          },
          "system_ram_bytes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override the detected host RAM in bytes. Mostly useful in tests.",
            "title": "System Ram Bytes"
          },
          "vram_bytes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override the detected GPU VRAM in bytes. Used by the Stage F smoke script to simulate a 24/20/16 GB card on a real 32 GB 5090.",
            "title": "Vram Bytes"
          }
        },
        "title": "AutoTuneVramRequest",
        "type": "object"
      },
      "AutoTuneVramResponse": {
        "description": "One row of the feasibility table returned to the Training UI.\n\nMirrors ``training_worker.engine.gpu_budget.LowVramRecommendation``\none-to-one. The frontend renders ``tier_label`` and ``warning``\nverbatim and binds the three knob fields to the start-job form.",
        "properties": {
          "blocks_resident_on_gpu": {
            "title": "Blocks Resident On Gpu",
            "type": "integer"
          },
          "confidence": {
            "enum": [
              "baseline",
              "supported",
              "plausible",
              "unsupported"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "detected_system_ram_bytes": {
            "title": "Detected System Ram Bytes",
            "type": "integer"
          },
          "detected_vram_bytes": {
            "title": "Detected Vram Bytes",
            "type": "integer"
          },
          "estimated_peak_vram_gb": {
            "title": "Estimated Peak Vram Gb",
            "type": "number"
          },
          "estimated_throughput_multiplier": {
            "title": "Estimated Throughput Multiplier",
            "type": "number"
          },
          "gradient_checkpointing": {
            "title": "Gradient Checkpointing",
            "type": "boolean"
          },
          "low_vram_mode": {
            "enum": [
              "off",
              "fp8",
              "nf4"
            ],
            "title": "Low Vram Mode",
            "type": "string"
          },
          "required_host_ram_gb": {
            "title": "Required Host Ram Gb",
            "type": "integer"
          },
          "tier_label": {
            "title": "Tier Label",
            "type": "string"
          },
          "warning": {
            "default": "",
            "title": "Warning",
            "type": "string"
          }
        },
        "required": [
          "tier_label",
          "low_vram_mode",
          "blocks_resident_on_gpu",
          "gradient_checkpointing",
          "estimated_peak_vram_gb",
          "estimated_throughput_multiplier",
          "required_host_ram_gb",
          "confidence",
          "detected_vram_bytes",
          "detected_system_ram_bytes"
        ],
        "title": "AutoTuneVramResponse",
        "type": "object"
      },
...
      "CheckpointInfo": {
        "description": "Info about a saved checkpoint.",
        "properties": {
          "epoch": {
            "title": "Epoch",
            "type": "integer"
          },
          "loss": {
            "title": "Loss",
            "type": "number"
          },
          "lr": {
            "title": "Lr",
            "type": "number"
          },
          "meta_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Path"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "step": {
            "title": "Step",
            "type": "integer"
          },
          "weights_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights Path"
          }
        },
        "required": [
          "step",
          "epoch",
          "loss",
          "lr"
        ],
        "title": "CheckpointInfo",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "RestartTrainingRequest": {
        "description": "Optional body for `POST /jobs/{job_id}/restart`.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "RestartTrainingRequest",
        "type": "object"
      },
...
      "StartTrainingRequest": {
        "description": "Request to start a new training job.",
        "properties": {
          "config_overrides": {
            "additionalProperties": true,
            "title": "Config Overrides",
            "type": "object"
          },
          "dataset_dir": {
            "title": "Dataset Dir",
            "type": "string"
          },
          "gpu_index": {
            "default": 0,
            "title": "Gpu Index",
            "type": "integer"
          },
          "model_path": {
            "default": "",
            "title": "Model Path",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "preset_id": {
            "default": "character_v1",
            "enum": [
              "character_v1",
              "concept_v1"
            ],
            "title": "Preset Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "trigger_word": {
            "default": "",
            "title": "Trigger Word",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "dataset_dir"
        ],
        "title": "StartTrainingRequest",
        "type": "object"
      },
...
      "TrainingJobSummary": {
        "description": "Brief summary of a job for list views.",
        "properties": {
          "created_at": {
            "default": 0.0,
            "title": "Created At",
            "type": "number"
          },
          "current_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Loss"
          },
          "current_step": {
            "title": "Current Step",
            "type": "integer"
          },
          "gpu_index": {
            "default": 0,
            "title": "Gpu Index",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "state": {
            "enum": [
              "created",
              "starting",
              "running",
              "paused",
              "completed",
              "cancelled",
              "errored"
            ],
            "title": "State",
            "type": "string"
          },
          "total_steps": {
            "title": "Total Steps",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "project_id",
          "state",
          "current_step",
          "total_steps"
        ],
        "title": "TrainingJobSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenLTX Trainer Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/training/auto-tune-vram": {
      "post": {
        "description": "Recommend a low-VRAM tier for the detected GPU + host RAM.\n\nStage F (per ``memory-bank/feature_real_training.md``). The\nresponse mirrors one row of the feasibility table and the\nTraining tab binds the three knob fields verbatim.",
        "operationId": "route_auto_tune_vram_api_training_auto_tune_vram_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AutoTuneVramRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoTuneVramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Auto Tune Vram",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs": {
      "get": {
        "operationId": "route_list_jobs_api_training_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrainingJobSummary"
                  },
                  "title": "Response Route List Jobs Api Training Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Jobs",
        "tags": [
          "training"
        ]
      },
      "post": {
        "operationId": "route_start_job_api_training_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Start Job",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}": {
      "delete": {
        "description": "Delete a terminal-state job. Errors if the job is still active.",
        "operationId": "route_delete_job_api_training_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Route Delete Job Api Training Jobs  Job Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Delete Job",
        "tags": [
          "training"
        ]
      },
      "get": {
        "operationId": "route_get_job_api_training_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TrainingJobRecord"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Route Get Job Api Training Jobs  Job Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Get Job",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "route_cancel_job_api_training_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Cancel Job",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/checkpoints": {
      "get": {
        "operationId": "route_list_checkpoints_api_training_jobs__job_id__checkpoints_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CheckpointInfo"
                  },
                  "title": "Response Route List Checkpoints Api Training Jobs  Job Id  Checkpoints Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Checkpoints",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/pause": {
      "post": {
        "operationId": "route_pause_job_api_training_jobs__job_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Pause Job",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/progress": {
      "get": {
        "operationId": "route_get_progress_api_training_jobs__job_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since_step",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Since Step",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingProgressSlice"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Get Progress",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/restart": {
      "post": {
        "description": "Spawn a new job using the same config as an existing one.",
        "operationId": "route_restart_job_api_training_jobs__job_id__restart_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RestartTrainingRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Restart Job",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/resume": {
      "post": {
        "operationId": "route_resume_job_api_training_jobs__job_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Resume Job",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/samples": {
      "get": {
        "operationId": "route_list_samples_api_training_jobs__job_id__samples_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SampleInfo"
                  },
                  "title": "Response Route List Samples Api Training Jobs  Job Id  Samples Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Samples",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/presets": {
      "get": {
        "operationId": "route_list_presets_api_training_presets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "title": "Response Route List Presets Api Training Presets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Presets",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/vram-sweep": {
      "get": {
        "description": "Return the full measured VRAM benchmark sweep.\n\nThe Training tab renders every measured (profile, quant,\nblocks_resident) cell as a sortable table so the operator can\npick any combination, not just the auto-tune recommendation.\nThe data is static (transcribed from the master sweep).",
        "operationId": "route_get_vram_sweep_api_training_vram_sweep_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VramSweepResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Get Vram Sweep",
        "tags": [
          "training"
        ]
      }
    },
...
  }
}